feat(file): add verify api & use FileProvider
