+++ officesdk/interface.go
@@ -2,13 +2,11 @@ package officesdk
 
 import "github.com/gin-gonic/gin"
 
-type PreviewProvider interface {
+type FileProvider interface {
 	GetFile(c *gin.Context, fileId string) (*FileResponse, error)
 	GetFileDownload(c *gin.Context, fileId string) (*DownloadResponse, error)
 	GetFileWatermark(c *gin.Context, fileId string) (*WatermarkResponse, error)
-}
 
-type EditProvider interface {
 	GetUploadURL(c *gin.Context, fileId string) (*UploadURLResponse, error)
 	CompleteUpload(c *gin.Context, fileId string) (*UploadCompletionResponse, error)
 	GetDownloadURL(c *gin.Context, fileId string) (*DownloadResponse, error)
@@ -16,6 +14,8 @@ type EditProvider interface {
 	GetAssetUploadURL(c *gin.Context, fileId string) (*AssetUploadURLResponse, error)
 	AssetCompleteUpload(c *gin.Context, fileId string) (*UploadCompletionResponse, error)
 	GetAssetDownloadURL(c *gin.Context, fileId string) (*DownloadResponse, error)
+
+	VerifyFile(c *gin.Context, fileId string) (*VerifyResponse, error)
 }
 
 type AIProvider interface {
